+++ example/lib/widgets/malware_bottom_sheet.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:freerasp/freerasp.dart';
+
+/// Bottom sheet widget that displays malware information
+class MalwareBottomSheet extends StatelessWidget {
+  /// Represents malware information in the example app
+  const MalwareBottomSheet({super.key, required this.suspiciousApps});
+
+  /// List of suspicious apps
+  final List<SuspiciousAppInfo> suspiciousApps;
+
+  @override
+  Widget build(BuildContext context) {
+    final textTheme = Theme.of(context).textTheme;
+
+    return Container(
+      padding: const EdgeInsets.all(16),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Text('Suspicious Apps', style: textTheme.titleMedium),
+          const SizedBox(height: 8),
+          ...suspiciousApps.map((malware) {
+            return ListTile(
+              title: Text(malware.packageInfo.packageName),
+              subtitle: Text('Reason: ${malware.reason}'),
+              leading: const Icon(Icons.warning, color: Colors.red),
+            );
+          }),
+          const SizedBox(height: 16),
+          SizedBox(
+            width: double.infinity,
+            child: FilledButton(
+              onPressed: () => Navigator.pop(context),
+              child: const Text('Dismiss'),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
