+++ example/lib/widgets/safety_icon.dart
